Finishing touches and reorganizing
