+++ proposals/templates/proposals/pdf/new_pdf_test.html
@@ -30,4 +30,16 @@
         {% endfor %}    
     {% endif %}
 
+    {% if extra_documents %}
+        {% for document in extra_documents %}
+            {% include document %}
+        {% endfor %}
+    {% endif %}
+
+    {% if dmp_file %}
+        {% include dmp_file %}
+    {% endif %}
+
+    {% include embargo %}
+
 {% endblock %}
