+++ proposals/templates/proposals/pdf/table_with_header.html
@@ -1,4 +1,4 @@
-{% if section_title is not None %}
+{% if section_title %}
     <h2> {{ section_title }} </h2>
 {% endif %}
 {% if study_title %}
