+++ proposals/utils/__init__.py
@@ -1,1 +1,2 @@
 from .proposal_utils import *
+from .pdf_diff_logic import *
