+++ proposals/utils/pdf_diff_logic.py
@@ -1,0 +1,779 @@
+
+from datetime import date
+from copy import copy
+
+from django.conf import settings
+from django.contrib.auth import get_user_model
+from django.template.loader import get_template
+from django.utils.html import format_html
+from django.utils.safestring import mark_safe
+from django.utils.translation import ugettext as _
+
+from proposals.templatetags.proposal_filters import needs_details, medical_traits, \
+necessity_required, has_adults
+
+class PDFSection:
+    
+    section_title = None
+    row_fields = None
+
+    def __init__(self, object):
+        self.object = object
+        # Create a copy of the class level row_fields, such that we can safely manipulate it without changing the class value
+        self._row_fields = copy(self.row_fields)
+
+    def render(self, context):
+        context = context.flatten()
+        template = get_template("proposals/pdf/table_with_header.html")
+        context.update(
+            {
+                "section_title": self.section_title,
+                "rows": self.get_rows(),
+            }
+        )
+        return template.render(context)
+    
+    def get_rows(self):
+
+        rows = [RowClass(self.object, field) for field in self._row_fields]
+
+        return rows
+        
+    def get_study_title(self, study):
+        if study.name:
+            study_title = format_html('{}{}{}{}{}',
+                                      _('Traject '),
+                                      study.order,
+                                      mark_safe(' <em>'),
+                                      study.name,
+                                      mark_safe(' </em>')
+            )
+        else:
+            study_title = format_html('{}{}',
+                                      _('Traject'),
+                                      {study.order}
+            )
+        return study_title
+    
+    def get_session_title(self, session):
+
+        order = session.order 
+        study_order = session.study.order 
+        study_name = session.study.name 
+        studies_number = session.study.proposal.studies_number 
+        sessions_number = session.study.sessions_number
+
+        if studies_number > 1 and sessions_number > 1:
+            session_title = format_html('{}{}{}{}{}{}{}',
+                                      _('Traject '),
+                                      study_order,
+                                      mark_safe(' <em>'),
+                                      study_name,
+                                      mark_safe(' </em>, '),
+                                      _('sessie '),
+                                      order
+            )
+        elif studies_number > 1:
+            session_title = format_html('{}{}{}{}{}',
+                                      _('Traject '),
+                                      study_order,
+                                      mark_safe(' <em>'),
+                                      study_name,
+                                      mark_safe(' </em>')
+            )
+        elif sessions_number >= 1:
+            session_title = format_html('{}{}',
+                                        _('Sessie '),
+                                        order
+            )
+        return session_title
+    
+    def get_task_title(self, task):
+        order=task.order 
+        session_order=task.session.order 
+        study_order=task.session.study.order 
+        study_name=task.session.study.name 
+        studies_number=task.session.study.proposal.studies_number
+        if studies_number > 1:
+            task_title = format_html('{}{}{}{}{}{}{}{}{}',
+                                      _('Traject '),
+                                      study_order,
+                                      mark_safe(' <em>'),
+                                      study_name,
+                                      mark_safe(' </em>, '),
+                                      _('sessie '),
+                                      session_order,
+                                      _(', taak '),
+                                      order
+            )
+        else:
+            task_title = format_html('{}{}{}{}',
+                                     _('Sessie '),
+                                     session_order,
+                                     _(', taak '),
+                                     order
+                                     )
+        return task_title
+
+class GeneralSection(PDFSection):
+    '''This class generates the data for the general section of 
+    the PDF page.'''
+
+    section_title = _("Algemene informatie over de aanvraag")
+    row_fields = [
+        'relation',
+        'supervisor',
+        'student_program',
+        'student_context',
+        'student_context_details',
+        'student_justification',
+        'other_applicants',
+        'applicants',
+        'other_stakeholders',
+        'stakeholders',
+        'date_start',
+        'title',
+        'funding',
+        'funding_details',
+        'funding_name',
+        'self_assessment',
+    ]
+    
+    def get_rows(self):
+        obj = self.object
+        rows = self._row_fields
+
+        if not obj.relation.needs_supervisor:
+            rows.remove('supervisor')
+        if not obj.relation.check_in_course:
+            rows.remove('student_program')
+            rows.remove('student_context')
+            if obj.student_context is not None:
+                if not obj.student_context.needs_details:
+                    rows.remove('student_context_details')
+            else:
+                rows.remove('student_context_details')
+            rows.remove('student_justification')
+        if not obj.other_applicants:
+            rows.remove('applicants')
+        if not obj.other_stakeholders:
+            rows.remove('stakeholders')
+        if not needs_details(obj.funding.all()):
+            rows.remove('funding_details')
+        if not needs_details(obj.funding.all(), 'needs_name'):
+            rows.remove('funding_name')
+
+        rows = [RowClass(obj, field) for field in rows]
+
+        return rows
+    
+class WMOSection(PDFSection):
+    '''Object for this section is proposal.wmo'''
+    section_title = _("Ethische toetsing nodig door een Medische Ethische Toetsingscommissie (METC)?")
+    row_fields = [
+        'get_metc_display',
+        'metc_details',
+        'metc_institution',
+        'get_is_medical_display',  
+    ]
+    
+    def get_rows(self):
+        obj = self.object
+        rows = self._row_fields
+
+        if not obj.metc == 'Y':
+            rows.remove('metc_details')
+            rows.remove('metc_institution')
+        else:
+            rows.remove('get_is_medical_display')
+
+        rows = [RowClass(obj, field) for field in rows]
+
+        return rows
+    
+class METCSection(PDFSection):
+    '''Object for this section is proposal.wmo
+    This class exists because the RowValueClass does some
+    funky things for working with the metc_decision_pdf field'''
+    section_title = _("Aanmelding bij de METC")
+    
+    row_fields = [
+        'metc_application',
+        'metc_decision',
+        'metc_decision_pdf'
+    ]
+    
+class TrajectoriesSection(PDFSection):
+
+    section_title = _("Eén of meerdere trajecten?")
+
+    row_fields = [
+        'studies_similar',
+        'studies_number'
+    ]
+    
+    def get_rows(self):
+        obj = self.object
+        rows = self._row_fields
+
+        if obj.studies_similar:
+            rows.remove('studies_number')
+
+        rows = [RowClass(obj, field) for field in rows]
+
+        return rows
+    
+class StudySection(PDFSection):
+    '''object for this study is proposal.study'''
+    section_title = _('De Deelnemers')
+    row_fields = [
+        'age_groups',
+        'legally_incapable',
+        'legally_incapable_details',
+        'has_special_details',
+        'special_details',
+        'traits',
+        'traits_details',
+        'necessity',
+        'necessity_reason',
+        'recruitment',
+        'recruitment_details',
+        'compensation',
+        'compensation_details',
+        'hierarchy',
+        'hierarchy_details',
+    ]
+
+    def get_rows(self):
+        obj = self.object
+        rows = self._row_fields
+
+        if not has_adults(obj):
+            rows.remove('legally_incapable')
+            rows.remove('legally_incapable_details')
+        elif not obj.legally_incapable:
+            rows.remove('legally_incapable_details')
+        if not obj.has_special_details:
+            rows.remove('special_details')
+            rows.remove('traits')
+            rows.remove('traits_details')
+        elif not medical_traits(obj.special_details.all()):
+            rows.remove('traits')
+            rows.remove('traits_details')
+        elif not needs_details(obj.traits.all()):
+            rows.remove('traits_details')
+        if not necessity_required(obj):
+            rows.remove('necessity')
+            rows.remove('necessity_reason')
+        if not needs_details(obj.recruitment.all()):
+            rows.remove('recruitment_details')
+        if not obj.compensation.needs_details:
+            rows.remove('compensation_details')
+        if not obj.hierarchy:
+            rows.remove('hierarchy_details')
+
+        rows = [RowClass(obj, field) for field in rows]
+
+        return rows
+        
+    def render(self, context):
+        if self.object.proposal.studies_number > 1:
+            context.update(
+                {
+                    'study_title': super().get_study_title(self.object)
+                }
+            )
+        return super().render(context)
+
+class InterventionSection(PDFSection):
+    '''This class will receive a intervention object'''
+    section_title = _('Het interventieonderzoek')
+    row_fields = [
+        'setting',
+        'setting_details',
+        'supervision',
+        'leader_has_coc',
+        'period',
+        'multiple_sessions', 
+        'session_frequency',
+        'amount_per_week',
+        'duration',
+        'measurement',
+        'experimenter',
+        'description',
+        'has_controls',
+        'controls_description',
+        'extra_task'
+    ]
+
+    def get_rows(self):
+        obj = self.object
+        rows = self._row_fields
+
+        if obj.version == 1:
+            fields_to_remove = ['multiple_sessions',
+                                'session_frequency',
+                                'extra_task']
+            for field in fields_to_remove:
+                rows.remove(field)
+        else:
+            rows.remove('amount_per_week')
+            if not obj.multiple_sessions:
+                rows.remove('session_frequency')
+            if obj.settings_contains_schools:
+                rows.remove('extra_task')
+        
+        if not needs_details(obj.setting.all()):
+            rows.remove('setting_details')
+        if not obj.study.has_children() or \
+        not needs_details(obj.setting.all(), 'needs_supervision'):
+            rows.remove('supervision')
+            rows.remove('leader_has_coc')
+        elif obj.supervision:
+            rows.remove('leader_has_coc')
+        if not obj.has_controls:
+            rows.remove('controls_description')   
+
+        rows = [RowClass(obj, field) for field in rows]
+
+        return rows
+    
+    def render(self, context):
+        if self.object.study.proposal.studies_number > 1:
+            context.update(
+                {
+                    'study_title': super().get_study_title(self.object.study)
+                }
+            )
+        return super().render(context)
+    
+class ObservationSection(InterventionSection):
+    '''Gets passed an observation object'''
+    section_title = _('Het observatieonderzoek')
+    row_fields = [
+        'setting',
+        'setting_details',
+        'supervision',
+        'leader_has_coc',
+        'days',
+        'mean_hours',
+        'details_who',
+        'details_why',
+        'details_frequency',
+        'is_anonymous',
+        'is_anonymous_details',
+        'is_in_target_group',
+        'is_in_target_group_details',
+        'is_nonpublic_space',
+        'is_nonpublic_space_details',
+        'has_advanced_consent',
+        'has_advanced_consent_details',
+        'needs_approval',
+        'approval_institution',
+        'approval_document',
+        'registrations',
+        'registrations_details'        
+    ]
+
+    def get_rows(self):
+        obj = self.object
+        rows = self._row_fields
+
+        if obj.version == 1:
+            to_remove_if_v1 = ['details_who',
+                            'details_why',
+                            'is_anonymous_details', 
+                            'is_in_target_group_details',
+                            'is_nonpublic_space_details',
+                            'has_advanced_consent_details'
+                            ]
+            for field in to_remove_if_v1:
+                rows.remove(field)
+
+            if not obj.is_nonpublic_space:
+                rows.remove('has_advanced_consent')
+            if not obj.needs_approval:
+                rows.remove('approval_institution')
+                rows.remove('approval_document')
+            elif obj.study.proposal.is_practice():
+                rows.remove('approval_document')
+        else:
+            to_remove_if_v2 = ['days', 'mean_hours', 'approval_document']
+            for field in to_remove_if_v2:
+                rows.remove(field)
+
+            if not obj.is_anonymous:
+                rows.remove('is_anonymous_details')
+            if not obj.is_in_target_group:
+                rows.remove('is_in_target_group_details')
+            if not obj.is_nonpublic_space:
+                rows.remove('is_nonpublic_space_details')
+                rows.remove('has_advanced_consent')
+                rows.remove('has_advanced_consent_details')
+            elif obj.has_advanced_consent:
+                rows.remove('has_advanced_consent_details')
+            if not needs_details(obj.setting.all(), 'is_school'):
+                rows.remove('needs_approval')
+            if not obj.needs_approval:
+                rows.remove('approval_institution')
+
+        if not needs_details(obj.setting.all()):
+            rows.remove('setting_details')
+        if not obj.study.has_children() or \
+        not needs_details(obj.setting.all(), 'needs_supervision'):
+            rows.remove('supervision')
+            rows.remove('leader_has_coc')
+        elif obj.supervision:
+            rows.remove('leader_has_coc')
+        if not needs_details(obj.registrations.all()):
+            rows.remove('registrations_details')
+        
+        rows = [RowClass(obj, field) for field in rows]
+
+        return rows
+
+class SessionsSection(StudySection):
+    '''Gets passed a study object
+    This Section looks maybe a bit unnecessary, but it does remove some logic, plus
+    the study_title.html from the template.'''
+
+    section_title = _("Het takenonderzoek en interviews")
+
+    def get_rows(self):
+
+        rows = [RowValueClass(self.object, 'sessions_number')]
+
+        return rows
+    
+class SessionSection(PDFSection):
+    '''Gets passed a session object'''
+    
+    row_fields = [
+        'setting',
+        'setting_details',
+        'supervision',
+        'leader_has_coc',
+        'tasks_number',       
+    ]
+
+    def get_rows(self):
+        obj = self.object
+        rows = self._row_fields
+
+        if not needs_details(obj.setting.all()):
+            rows.remove('setting_details')
+        if not obj.study.has_children() or \
+        not needs_details(obj.setting.all(), 'needs_supervision'):
+            rows.remove('supervision')
+            rows.remove('leader_has_coc')
+        elif obj.supervision:
+            rows.remove('leader_has_coc')
+        
+        rows = [RowClass(obj, field) for field in rows]
+
+        return rows
+    
+    def render(self, context):
+        context.update(
+            {
+                'study_title': super().get_session_title(self.object)
+            }
+        )
+        return super().render(context)
+    
+class TaskSection(PDFSection):
+    '''Gets passed a task object'''
+    
+    row_fields = [
+        'name',
+        'duration',
+        'registrations',
+        'registrations_details',
+        'registration_kinds',
+        'registration_kinds_details',
+        'feedback',
+        'feedback_details',
+        'description'
+    ]
+
+    def get_rows(self):
+        obj = self.object
+        rows = self._row_fields
+
+        if not needs_details(obj.registrations.all()):
+            rows.remove('registrations_details')
+        if not needs_details(obj.registrations.all(), 'needs_kind') or \
+        not needs_details(obj.registration_kinds.all()):
+            rows.remove('registration_kinds')
+            rows.remove('registration_kinds_details')
+        elif not needs_details(obj.registration_kinds.all()):
+            rows.remove('registration_kinds_details')
+        if not obj.feedback:
+            rows.remove('feedback_details')
+        
+        rows = [RowClass(obj, field) for field in rows]
+
+        return rows
+    
+    def render(self, context):
+        context.update(
+            {
+                'study_title': super().get_task_title(self.object)
+            }
+        )
+        return super().render(context)
+    
+class TasksOverviewSection(PDFSection):
+    '''Gets passed a session object
+    This might be unecessary ... Maybe just use the existing template language ...
+    Because the verbose name of this field is formatted, the method for retrieving
+    the verbose name in the RowClass is quite convoluded.'''
+
+    row_fields = [
+        'tasks_duration'
+    ]
+
+    def render(self, context):
+        '''Here I am just overriding the render function, to get the correct formatting
+        The naming is a bit confusing, might fix later'''
+        context.update(
+            {
+                'study_title': _('Overzicht van het takenonderzoek')
+            }
+        )
+        return super().render(context)
+    
+class StudyOverviewSection(StudySection):
+    '''Receives a Study object'''
+
+    section_title = _('Overzicht en eigen beoordeling van het gehele onderzoek')
+    row_fields = [
+        'deception',
+        'deception_details',
+        'negativity',
+        'negativity_details',
+        'stressful',
+        'stressful_details',
+        'risk',
+        'risk_details'
+    ]
+
+    def get_rows(self):
+        obj = self.object
+        rows = self._row_fields
+
+        rows_to_remove = []
+        for x in range(0, 7, 2):
+            if getattr(obj, rows[x]) == 'N':
+                rows_to_remove.append(rows[x+1])
+        rows = [row for row in rows if row not in rows_to_remove]
+
+        if not obj.has_sessions and not obj.deception == 'N':
+            rows.remove('deception')
+            rows.remove('deception_details')
+        elif not obj.has_sessions:
+            rows.remove('deception')
+        
+        rows = [RowClass(obj, field) for field in rows]
+
+        return rows
+    
+class InformedConsentFormsSection(InterventionSection):
+    '''Receives a Documents object'''
+
+    section_title = _('Informed consent formulieren')
+
+    def get_rows(self):
+        obj = self.object
+
+        rows = []
+
+        rows.append(RowClass(obj.proposal, 'translated_forms'))
+        if obj.proposal.translated_forms:
+            rows.append(RowClass(obj.proposal, 'translated_forms_languages'))
+        if not obj.proposal.is_practice() and obj.informed_consent:
+            rows.append(RowClass(obj, 'informed_consent'))
+            rows.append(RowClass(obj, 'briefing'))
+        if obj.study.passive_consent is not None:
+            rows.append(RowClass(obj.study, 'passive_consent'))
+        if obj.study.passive_consent:
+            rows.append(RowClass(obj.study, 'passive_consent_details'))
+        if obj.director_consent_declaration:
+            rows.append(RowClass(obj, 'director_consent_declaration'))
+        if obj.director_consent_information:
+            rows.append(RowClass(obj, 'director_consent_information')) 
+        if obj.parents_information:
+            rows.append(RowClass(obj, 'parents_information'))
+
+        return rows
+    
+class ExtraDocumentsSection(PDFSection):
+    '''gets a documents object'''
+
+    row_fields = [
+        'informed_consent',
+        'briefing'
+    ]
+
+    def __init__(self, object, count):
+        super().__init__(object)
+        self.section_title = _('Extra formulieren ') + str(count)
+    
+    def get_rows(self):
+        obj = self.object
+        rows = self._row_fields
+
+        if not obj.informed_consent:
+            rows.remove('informed_consent')
+        if not obj.briefing:
+            rows.remove('briefing')
+        
+        rows = [RowClass(obj, field) for field in rows]
+
+        return rows
+    
+class DMPFileSection(PDFSection):
+    '''Gets passed a proposal object.'''
+
+    section_title = _('Data Management Plan')
+    
+    row_fields = ['dmp_file']
+
+class EmbargoSection(PDFSection):
+    '''Gets passed a proposal object'''
+
+    section_title = _('Aanmelding Versturen')
+
+    row_fields = [
+        'embargo',
+        'embargo_end_date'
+    ]
+
+    def get_rows(self):
+        obj = self.object
+        rows = self._row_fields
+
+        if not obj.embargo:
+            rows.remove('embargo_end_date')
+        
+        rows = [RowClass(obj, field) for field in rows]
+
+        return rows
+
+class RowClass:
+
+    verbose_name_diff_field_dict = {
+        'get_metc_display': 'metc',
+        'get_is_medical_display': 'is_medical'
+    }
+
+    def __init__(self, object, field):
+        self.object = object
+        self.field = field
+    
+    def verbose_name(self):
+        if self.field in self.verbose_name_diff_field_dict:
+            verbose_name_field = self.verbose_name_diff_field_dict[self.field]
+            verbose_name = self.get_verbose_name(verbose_name_field)
+        else:
+            verbose_name= self.get_verbose_name(self.field)
+        return verbose_name   
+    
+    def value(self):
+        return RowValueClass(self.object, self.field).render()
+    
+    def get_verbose_name(self, field):
+        if field != 'tasks_duration':
+            return mark_safe(self.object._meta.get_field(field).verbose_name)
+        else:
+            return mark_safe(self.object._meta.get_field(field).verbose_name % self.object.net_duration())
+
+class RowValueClass:
+
+    def __init__(self, object, field):
+
+        self.object = object
+        self.field = field
+
+    def render(self):
+        from ..models import Funding, Relation
+
+
+        value = getattr(self.object, self.field)
+
+        User = get_user_model()
+
+        if value in ('Y', 'N', '?'):
+            return self.yes_no_doubt(value)
+        if type(value) in (str, int, date):
+            return value
+        elif value is None:
+            return _('Onbekend')
+        elif type(value) == bool:
+            return _('Ja') if value else _('Nee')
+        elif type(value) == User:
+            return self.handle_user(value)
+        elif type(value) == Relation:
+            return value.description
+        elif value.__class__.__name__ == 'ManyRelatedManager':
+            if value.all().model == User:
+                return self.get_applicants_names(value)
+            else:
+                return self.get_object_list(value)
+        elif value.__class__.__name__ == 'FieldFile':
+            return self.handle_field_file(value, self.object)
+        elif callable(value):
+            return value()
+        
+    def handle_user(self, user):
+        return user.get_full_name()
+    
+    def get_applicants_names(self, applicants):
+        applicant_names = [applicant.get_full_name() for applicant in applicants.all()]
+        return self.create_unordered_html_list(applicant_names)
+    
+    def get_object_list(self, object):
+        list_of_objects = [obj for obj in object.all()]
+        return self.create_unordered_html_list(list_of_objects)
+    
+    def create_unordered_html_list(self, list):
+        html_output = mark_safe('<ul class="p-0">')
+
+        for item in list:
+            html_output += format_html('{}{}{}',
+                mark_safe('<li>'),
+                item,
+                mark_safe('</li>')                
+            )
+        
+        html_output += mark_safe('</ul>')
+
+        return html_output
+    
+    def handle_field_file(self, field_file, object):
+        from ..models import Proposal
+        if object == Proposal and field_file == object.wmo.metc_decision_pdf:
+            if object.wmo.metc_decision_pdf and not object.is_practice():
+                output = format_html('{}{}{}{}{}',
+                                    mark_safe('<a href="'),
+                                    f'{settings.BASE_URL}{field_file.url}',
+                                    mark_safe('" target="_blank">'),
+                                    _('Download'),
+                                    mark_safe('</a>')
+                                    )
+            else:
+                output = _('Niet aangeleverd')
+        else:
+            output = format_html('{}{}{}{}{}',
+                                    mark_safe('<a href="'),
+                                    f'{settings.BASE_URL}{field_file.url}',
+                                    mark_safe('" target="_blank">'),
+                                    _('Download'),
+                                    mark_safe('</a>')
+                                    )
+        return output
+    
+    def yes_no_doubt(self, value):
+        from main.models import YES_NO_DOUBT
+        d = dict(YES_NO_DOUBT)
+        return d[value]
